Introduce Conditional RGYB
Just show buttons with code in use.
CONDITIONAL_RGYB will only show both buttons and used codes those that are not in use are ignored.

This is what I was trying to explain in its day with RAED, as this is our skin, by default, we will make it elegant.