Revised: MessageBox and MessageBoxSimple
the skin is now ready to accept both icon and Multipixmap as well as accept the usual Pixmap.

disabled self.instance.resize (eSize (* wsize)):
if we use this code it will always take the form of the Messagebox screen ignoring MessageBoxSimple and it must be taken into account that in the flash slots for boxes of this type of partitions we need more width for the user.