Revised: ChannelSelection
Up to 12 EPG events have been added to the channel list.
EPGImport has been revised.
Added filtersServicesSetup and filtersServicesSelection

The list of channels are the most used screens, with service next events we have an average  of seven eight hours of EPG, the objective is that the user does not have to go to another screen to search for a Single EPG, if not to receive this information from the same channel list.