Remove unknown parameters
When this skin was installed, these parameters were by default.
We are not going to leave a single parameter that we really do not have to modify from our E2 ... and if the E2 code requires it, then we will import it.
I leave by default only those that I know and I know are needed regarding our skin and E2-OV, the rest alone does give a headache.